feat: add HomePage component
