+++ src/App.jsx
@@ -1,13 +1,10 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
 import { hot } from 'react-hot-loader';
 import CssBaseline from '@material-ui/core/CssBaseline';
 
-import MovieList from './constainers/MovieList';
-import Header from './constainers/Header';
-import Footer from './constainers/Footer';
+import HomePage from './pages/homePage';
 import ErrorBoundary from './components/ErrorBoundary';
-import moviesList from './utils/data';
 
 const theme = createMuiTheme({
   palette: {
@@ -32,53 +29,12 @@ const theme = createMuiTheme({
 });
 
 function App() {
-  const [movies, setMovies] = useState(moviesList);
-  const handleSort = sortBy => {
-    const sortedMovies = [
-      ...movies.sort((a, b) => {
-        const valueA = a[sortBy].toUpperCase();
-        const valueB = b[sortBy].toUpperCase();
-        if (valueA < valueB) {
-          return -1;
-        }
-        if (valueA > valueB) {
-          return 1;
-        }
-        return 0;
-      })
-    ];
-    setMovies(sortedMovies);
-  };
-  const handleSearch = query => {
-    if (query.trim() === '') {
-      setMovies(moviesList);
-      return;
-    }
-    const searchResults = moviesList.filter(movie =>
-      movie.title.toUpperCase().includes(query.trim().toUpperCase())
-    );
-    setMovies(searchResults);
-  };
-  const handleFilter = filter => {
-    if (filter === 'all') {
-      setMovies(moviesList);
-      return;
-    }
-    const filteredList = moviesList.filter(movie => {
-      const formatedGenres = movie.genres.map(genre => genre.toUpperCase());
-      return formatedGenres.includes(filter.toUpperCase());
-    });
-    setMovies(filteredList);
-  };
-
   return (
     <>
       <ThemeProvider theme={theme}>
         <CssBaseline />
         <ErrorBoundary>
-          <Header handleSearch={handleSearch} />
-          <MovieList movies={movies} handleSort={handleSort} handleFilter={handleFilter} />
-          <Footer />
+          <HomePage />
         </ErrorBoundary>
       </ThemeProvider>
     </>
